+++ packages/rsi-api-base/src/index.ts
@@ -1,5 +1,5 @@
 import { EventEmitter } from 'events';
-
+/* const axios = require('axios') */
 export default class Base {
   emitter: any;
 
@@ -7,64 +7,29 @@ export default class Base {
     this.emitter = new EventEmitter();
   }
 
-  gatewayRequest(apiKey: string, roomName: string) {
+  async gatewayRequest(apiKey: string, roomName: string) {
     if (!apiKey) {
       throw Error('base akadu-rsi: apiKey is undefined.');
     }
     if (!roomName) {
-      throw Error('base akkadu-rsi: roomname is undefined.');
+      throw Error('base akkadu-rsi: roomName is undefined.');
     }
     // we get the stream /events/{eventId}/streams 
-    const streamTest = {
-      "auth": {
-        "uid": 0,
-        "channel": "string",
-        "appId": "string",
-        "token": "string",
-        "rtmToken": "string"
-      },
-      "publishers": [
-        {
-          "id": 0,
-          "userId": 0,
-          "eventId": 0,
-          "pairId": "string",
-          "interpreterId": 0,
-          "language": "string",
-          "sourceLanguage": true,
-          "sourceLanguageId": 0,
-          "interpreterLevel": 0,
-          "interpreterNeeded": true,
-          "createdAt": "2021-05-25T07:49:48.840Z",
-          "updatedAt": "2021-05-25T07:49:48.840Z"
-        }
-      ],
-      "floorLang": "en-US",
-      "userType": "string"
+
+    // the request below are for test purpose and will have to be removed 
+    // once the gateway set up.
+    const eventRequest = await(await fetch(`https://devapi.akkadu.com/v2/events/?roomName=${roomName}`)).json()
+    const eventId = eventRequest?.data?.events?.[0]?.id;
+    if(!eventId){
+      throw Error('no eventId for this event');
     }
-    const eventTest = {
-      bio: null,
-      countryCode: null,
-      createdAt: "2020-10-09T04:07:43.879Z",
-      endDate: "2021-05-13T15:07:00.000Z",
-      eventUrl: null,
-      id: 2545,
-      isActive: true,
-      isAkkaduEvent: true,
-      isPrivate: false,
-      language: null,
-      location: null,
-      name: "testApi2",
-      pdfUrl: null,
-      posterUrl: null,
-      roomName: "pkmj",
-      startDate: "2021-05-12T21:07:00.000Z",
-      tagIds: [4],
-      timeZone: null,
-      updatedAt: "2021-05-13T04:52:47.934Z",
-      userId: 622
-     }
-    return streamTest
+    const streamRequest = await (await fetch(`https://devapi.akkadu.com/v2/events/${eventId}/streams`)).json()
+    const eventLanguagesRequest = await  (await (fetch(`https://devapi.akkadu.com/v2/events/${eventId}/languages`))).json()
+    const eventLanguageState = await  (await (fetch(`https://devapi.akkadu.com/v2/language-state?roomname=${roomName}`))).json()
+    const stream = streamRequest?.data
+    const eventLanguages  = eventLanguagesRequest?.data.languages
+    const languageState = eventLanguageState?.data.languageState;
+    return { stream, languageState, eventLanguages }
   }
   on(event:any,fn:any) {
     this.emitter.on(event,fn)