+++ packages/rsi-api-interpretation-player/rollup.config.js
@@ -1,0 +1,26 @@
+import resolve from 'rollup-plugin-node-resolve'
+import commonjs from 'rollup-plugin-commonjs'
+import builtins from 'rollup-plugin-node-builtins'
+
+
+// We need to create a bundle of our rtc-streamer-consumer before importing it to typescript
+
+export default {
+  input: './streamers/index.js',
+  cache:false,
+  output: {
+    dir: './dist/',
+    format: 'esm',
+    name: 'bundle',
+    inlineDynamicImports:true,
+  },
+  plugins: [
+    resolve({
+      main:true
+    }),
+    commonjs(),
+    builtins()
+  ],
+  external: [
+  ],
+}