+++ packages/rsi-api-interpretation-player/streamers/index.js
@@ -1,0 +1,2 @@
+import Akkadu from '@akkadu/rtc-streamer-consumer' 
+export default Akkadu