+++ packages/rsi-api-react-example/src/App.js
@@ -3,7 +3,15 @@ import './App.css';
 import { RsiApiInterpretationPlayer, /* RsiApiInterpretationManager */ } from '@akkadu/rsi-api-react'
 import '@akkadu/rsi-api-react/dist/index.css'
 
-
+// We redirect to an URL with a rsi-roomname parameter (test-purposes)
+const params = new URLSearchParams(window.location.search)
+const roomname = params.get('rsi-roomname');
+var url = window.location.href
+var arr = url.split("/");
+var currentUrl = arr[0] + "//" + arr[2]
+if(!roomname){
+    window.location.href = `${currentUrl}/?rsi-roomname=pvzj`
+}
 const handleOnReady = (e) => {
   console.info('receive onReady event:',e );
 }