+++ packages/rsi-api-vue-example/src/views/home.vue
@@ -12,7 +12,18 @@ import InterpretationManager from '@/components/interpretation-manager.vue'
     components:{
       InterpretationPlayer,
       InterpretationManager 
-    }
+    },
+    mounted() {
+      // We redirect to an URL with a rsi-roomname parameter (test-purposes)
+       const params = new URLSearchParams(window.location.search)
+       const roomname = params.get('rsi-roomname');
+       var url = window.location.href
+       var arr = url.split("/");
+       var currentUrl = arr[0] + "//" + arr[2]
+       if(!roomname){
+          window.location.href = `${currentUrl}/?rsi-roomname=pvzj`
+       }
+    },
   }
 </script>