+++ packages/rsi-api-interpretation-player/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.11](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-api-interpretation-player@0.3.10...@akkadu/rsi-api-interpretation-player@0.3.11) (2021-06-04)
+
+**Note:** Version bump only for package @akkadu/rsi-api-interpretation-player
+
+
+
+
+
+## [0.3.10](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-api-interpretation-player@0.3.9...@akkadu/rsi-api-interpretation-player@0.3.10) (2021-06-03)
+
+**Note:** Version bump only for package @akkadu/rsi-api-interpretation-player
+
+
+
+
+
+## [0.3.9](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-api-interpretation-player@0.3.8...@akkadu/rsi-api-interpretation-player@0.3.9) (2021-06-03)
+
+**Note:** Version bump only for package @akkadu/rsi-api-interpretation-player
+
+
+
+
+
+## [0.3.8](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-api-interpretation-player@0.3.7...@akkadu/rsi-api-interpretation-player@0.3.8) (2021-06-03)
+
+**Note:** Version bump only for package @akkadu/rsi-api-interpretation-player
+
+
+
+
+
+## [0.3.7](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-api-interpretation-player@0.3.6...@akkadu/rsi-api-interpretation-player@0.3.7) (2021-06-03)
+
+**Note:** Version bump only for package @akkadu/rsi-api-interpretation-player
+
+
+
+
+
 ## [0.3.6](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-api-interpretation-player@0.3.5...@akkadu/rsi-api-interpretation-player@0.3.6) (2021-05-27)
 
 **Note:** Version bump only for package @akkadu/rsi-api-interpretation-player