+++ packages/rsi-api-react-example/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.10](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-api-react-example@0.1.9...@akkadu/rsi-api-react-example@0.1.10) (2021-06-04)
+
+**Note:** Version bump only for package @akkadu/rsi-api-react-example
+
+
+
+
+
+## [0.1.9](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-api-react-example@0.1.8...@akkadu/rsi-api-react-example@0.1.9) (2021-06-03)
+
+**Note:** Version bump only for package @akkadu/rsi-api-react-example
+
+
+
+
+
+## [0.1.8](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-api-react-example@0.1.7...@akkadu/rsi-api-react-example@0.1.8) (2021-06-03)
+
+**Note:** Version bump only for package @akkadu/rsi-api-react-example
+
+
+
+
+
+## [0.1.7](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-api-react-example@0.1.6...@akkadu/rsi-api-react-example@0.1.7) (2021-06-03)
+
+**Note:** Version bump only for package @akkadu/rsi-api-react-example
+
+
+
+
+
+## [0.1.6](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-api-react-example@0.1.5...@akkadu/rsi-api-react-example@0.1.6) (2021-06-03)
+
+**Note:** Version bump only for package @akkadu/rsi-api-react-example
+
+
+
+
+
 ## [0.1.5](https://github.com/Akkadu/rsi-api-widgets/compare/@akkadu/rsi-api-react-example@0.1.4...@akkadu/rsi-api-react-example@0.1.5) (2021-05-27)
 
 **Note:** Version bump only for package @akkadu/rsi-api-react-example